py_inference/requirements.txt
llama-cpp-python
pandas
matplotlib
seaborn
python-dotenv
